stop jOOQ narrowing every stored decimal to Float (#116) (#195)
jOOQ's SQLite dialect maps REAL to Float, so every stored decimal without an
exact 32-bit form was degraded on the round-trip: a 107.05 trend weight read
back 107.05000305175781. SQLite REAL is already an 8-byte double, so the
column type was never the problem — the type *mapping* was.

A single <forcedType> on REAL fixes it in one place, which then makes all 41
manual .toFloat()/.toDouble() conversions across the seven repositories
redundant. Zero manual conversions now remain in persistence: the DSL.sum()
BigDecimal hop went too, by letting jOOQ convert via fetchOne(0, Double).

Coverage is exact: 18 REAL columns generate 18 Double fields, no Float
remains, and the 15 Integer columns are untouched.

The tolerances this bug forced into the suite are tightened rather than a new
test being invented for its own sake — the asymmetry between 1e-2 on persisted
doubles and 1e-6 on computed ones *was* the bug. Reverting only the fix now
reddens 7 tests, including "a fresh Goal reads exactly 0%", which is ADR 0016's
own wording (line 28) that the narrowing had quietly falsified.

No schema migration, no API change (the OpenAPI spec regenerates byte-identical).
Rows written before this fix keep their narrowed value and are not backfilled;
reading them is unaffected, since float->double->float is identity and every
removed conversion was on the write side.

+++ backend/src/main/kotlin/com/tucker/persistence/GoalRepository.kt
@@ -23,9 +23,9 @@ class GoalRepository(private val dsl: DSLContext) {
     fun insert(goal: Goal): Goal {
         val rec = dsl.newRecord(GOAL)
         rec.startedOn = goal.startedOn.toString()
-        rec.startWeightKg = goal.startWeightKg.toFloat()
-        rec.targetWeightKg = goal.targetWeightKg.toFloat()
-        rec.rateKgPerWeek = goal.rateKgPerWeek.toFloat()
+        rec.startWeightKg = goal.startWeightKg
+        rec.targetWeightKg = goal.targetWeightKg
+        rec.rateKgPerWeek = goal.rateKgPerWeek
         rec.active = if (goal.active) 1 else 0
         rec.reachedOn = goal.reachedOn?.toString()
         rec.store()
@@ -48,9 +48,9 @@ class GoalRepository(private val dsl: DSLContext) {
     private fun GoalRecord.toGoal(): Goal = Goal(
         id = id!!.toLong(),
         startedOn = LocalDate.parse(startedOn),
-        startWeightKg = startWeightKg.toDouble(),
-        targetWeightKg = targetWeightKg.toDouble(),
-        rateKgPerWeek = rateKgPerWeek.toDouble(),
+        startWeightKg = startWeightKg,
+        targetWeightKg = targetWeightKg,
+        rateKgPerWeek = rateKgPerWeek,
         active = active != 0,
         reachedOn = reachedOn?.let(LocalDate::parse),
     )
